+++ src/main/java/com/monitora/preco/repository/ProdutoRepository.java
@@ -4,8 +4,6 @@ import com.monitora.preco.entity.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface ProdutoRepository extends JpaRepository<Produto, BigInteger> {
+public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
 }
